remapped_to/remapped_from should always refer to field names
Previously when FK columns were remapped, the user specified name was
used as the remapped_to value. This could be ambiguous. Instead the
remapped_to should point to the name of the remapped column, separate
from what the user specified as the column's name (that's set as the
field_display_name).